fix(compiler): move worker file to the root
